✨feat(console): add commands
