+++ src/app/components/index/index.component.html
@@ -67,7 +67,7 @@
 		</mat-accordion>
 	</perfect-scrollbar>
 </div>
-<div class="menu bottom" [class.accent]="(cadStatus | async)?.name !== 'normal'" [@closeBottom]="showBottomMenu ? 'open' : 'closed'">
+<div class="menu bottom" [class.accent]="cadStatusStr !== '普通'" [@closeBottom]="showBottomMenu ? 'open' : 'closed'">
 	<app-cad-console [cad]="cad" (afterOpenCad)="afterOpenCad()"></app-cad-console>
 	<div class="cad-status">当前状态：{{ cadStatusStr }}</div>
 </div>
