+++ src/app/components/menu/toolbar/toolbar.component.html
@@ -22,12 +22,12 @@
 	<button mat-menu-item (click)="fillet()">连接线段(圆角)</button>
 </mat-menu>
 <mat-menu #transform="matMenu">
-	<button mat-menu-item (click)="flip($event, true, false)">垂直翻转</button>
-	<button mat-menu-item (click)="flip($event, false, true)">水平翻转</button>
-	<button mat-menu-item (click)="flip($event, true, true)">中心对称</button>
-	<button mat-menu-item (click)="rotate($event, true)">顺时针旋转</button>
-	<button mat-menu-item (click)="rotate($event, false)">逆时针旋转</button>
-	<button mat-menu-item (click)="rotate($event)">旋转</button>
+	<button mat-menu-item (click)="flip(true, false)">垂直翻转</button>
+	<button mat-menu-item (click)="flip(false, true)">水平翻转</button>
+	<button mat-menu-item (click)="flip(true, true)">中心对称</button>
+	<button mat-menu-item (click)="rotate(true)">顺时针旋转</button>
+	<button mat-menu-item (click)="rotate(false)">逆时针旋转</button>
+	<button mat-menu-item (click)="rotate()">旋转</button>
 </mat-menu>
 
 <button mat-icon-button [matMenuTriggerFor]="view">视图</button>
