+++ src/app/components/message/message.component.scss
@@ -9,8 +9,8 @@ mat-form-field {
 	font-size: 1rem;
 }
 
-.mat-dialog-actions {
-	position: absolute;
-	bottom: 0;
-	width: 100%;
-}
+// .mat-dialog-actions {
+// 	position: absolute;
+// 	bottom: 0;
+// 	width: 100%;
+// }
